requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
atomicwrites==1.4.1
attrs==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.5
configparser==5.3.0
cycler==0.11.0
fonttools==4.28.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.4
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
plotly==5.4.0
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==7.1.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
tenacity==8.0.1
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.8
zipp==3.8.1
